Feat(api): Axios 초기 설정 관련 코드 통합 및 query 설정 추가 구현
