--- src/apis/index.ts
@@ -1,30 +1,0 @@
-import axios, { AxiosInstance, AxiosRequestConfig } from 'axios';
-
-import { BASE_URI } from '@/constants/URI';
-import { QueryClient } from '@tanstack/react-query';
-
-const initInstance = (config: AxiosRequestConfig): AxiosInstance => {
-  const instance = axios.create({
-    timeout: 1000,
-    ...config,
-    headers: {
-      'Content-Type': 'application/json',
-      'Cross-Control-Allow-Origin': '*',
-      ...config.headers,
-    },
-  });
-  return instance;
-};
-
-export const fetchInstance = initInstance({ baseURL: BASE_URI });
-
-export const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      retry: 3,
-      refetchOnMount: true,
-      refetchOnReconnect: true,
-      refetchOnWindowFocus: true,
-    },
-  },
-});
